Fix debugger tests
